+++ src/Enums/RegistrationMode.php
@@ -5,6 +5,5 @@ namespace NettSite\Messenger\Enums;
 enum RegistrationMode: string
 {
     case Open = 'open';
-    case Approval = 'approval';
     case Closed = 'closed';
 }