+++ src/Http/Requests/RegisterUserRequest.php
@@ -10,7 +10,7 @@ class RegisterUserRequest extends FormRequest
     {
         return [
             'name' => ['required', 'string', 'max:255'],
-            'email' => ['required', 'string', 'email', 'max:255', 'unique:messenger_users,email'],
+            'email' => ['required', 'string', 'email', 'max:255', 'unique:users,email'],
             'password' => ['required', 'string', 'min:8', 'confirmed'],
             'fcm_token' => ['nullable', 'string'],
             'platform' => ['required', 'string', 'in:android,ios,web'],